Got VAO generating per context but need to fix delegate call

+++ Editor/Core/Source/SceneView.cpp
@@ -18,11 +18,11 @@ bool VSceneView::Init()
 	VWindow::GetPrimaryMonitorSize(&mWidth, &mHeight);
 
 	//Work out positions
-	int x1 = mWidth * 0.25f, x2 = mWidth * 0.5f;
-	int y1 = 50, y2 = mHeight * 0.5f - 50;
+	float x1 = mWidth * 0.25f, x2 = mWidth * 0.5f;
+	float y1 = 50, y2 = mHeight * 0.5f - 50;
 
-	Window = CreateWindow(x2, y2, "Scene View", false);
-	Window->SetPosition(x1, y1);
+	Window = CreateWindow((int)x2, (int)y2, "Scene View", false);
+	Window->SetPosition((int)x1, (int)y1);
 
 	return true;
 }
